+++ libxcode/src/main/groovy/org/openbakery/xcode/Xcodebuild.groovy
@@ -169,8 +169,8 @@ class Xcodebuild {
 
 		commandList.add("CODE_SIGN_IDENTITY=")
 		commandList.add("CODE_SIGNING_REQUIRED=NO")
-		//commandList.add("CODE_SIGN_ENTITLEMENTS=")
-		//commandList.add("CODE_SIGNING_ALLOWED=NO")
+		commandList.add("CODE_SIGNING_ALLOWED=NO")
+
 	}
 
 	private boolean isSimulator() {