+++ plugin/src/test/groovy/org/openbakery/XcodeTestTaskSpecification.groovy
@@ -87,6 +87,7 @@ class XcodeTestTaskSpecification extends Specification {
 		return [
 						"CODE_SIGN_IDENTITY=",
 						"CODE_SIGNING_REQUIRED=NO",
+						"CODE_SIGNING_ALLOWED=NO"
 		]
 	}